Partially improve the readability of gpt plan generation
