+++ lib/gpt-plan-generate.ts
@@ -8,6 +8,7 @@ import {
 } from './google-map-api'
 import { min } from 'date-fns'
 import { getHistory, plusVisited } from './history-management'
+import { filterDistanceMatrixData, filterGoogleMapData } from './google-map-filter'
 
 const GPT_KEY = process.env.EXPO_PUBLIC_GPT_KEY
 export interface Activity {
@@ -31,72 +32,6 @@ export interface Plan {
   [key: number]: Activity[]
 }
 
-export interface GoogleMapPlace {
-  name?: string
-  vicinity?: string
-  rating?: number
-  user_ratings_total?: number
-  price_level?: number
-  types?: string[]
-  geometry?: {
-    location?: {
-      lat: number
-      lng: number
-    }
-  }
-  photos?: {
-    photo_reference?: string
-  }[]
-}
-
-export interface GoogleMapResponse {
-  results: GoogleMapPlace[]
-}
-
-export const filterGoogleMapData = (data: GoogleMapResponse, currentLocation: string) => {
-  try {
-    console.log("checking data")
-    if (!data || !data.results) {
-      return []
-    }
-    console.log("data", data.results)
-    const filteredResults = data.results
-      .filter((place) => {
-        return place.photos && place.photos.length > 0
-      })
-      .map((place: GoogleMapPlace) => ({
-        name: place.name || 'Unknown Name',
-        vicinity: place.vicinity || 'Unknown Vicinity',
-        rating: place.rating !== undefined ? place.rating : 0,
-        user_ratings_total:
-          place.user_ratings_total !== undefined ? place.user_ratings_total : 0,
-        price_level: place.price_level !== undefined ? place.price_level : 0,
-        types: place.types || [],
-        geometry: place.geometry?.location || { lat: 0, lng: 0 },
-        photo_reference: place.photos?.[0]?.photo_reference || '',
-        distance: fastDistance(parseFloat(currentLocation.split(',')[0]), parseFloat(currentLocation.split(',')[1]), place.geometry?.location?.lat || 0, place.geometry?.location?.lng || 0)
-      }))
-
-    return filteredResults
-  } catch (error) {
-    throw 'Error filtering Google Map data'
-  }
-}
-
-export const filterDistanceMatrixData = (data: any) => {
-  const filteredResults = data.rows.map((row: any, originIndex: number) => {
-    return row.elements.map((element: any, destinationIndex: number) => ({
-      origin: data.origin_addresses[originIndex] || 'Unknown Origin',
-      destination:
-        data.destination_addresses[destinationIndex] || 'Unknown Destination',
-      distance: element.distance?.text || '0 km',
-      duration: element.duration?.text || '0 mins',
-    }))
-  })
-
-  return filteredResults
-}
-
 const json_sample: Plan = {
   1: [
     {
@@ -114,44 +49,29 @@ const json_sample: Plan = {
       photo_reference: 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx',
       rating: 4.5,
       user_ratings_total: 2000
-    },
-    {
-      date: '2024-09-29',
-      time: '8:00',
-      duration: '10 min',
-      destination: 'X',
-      destinationDescrib: 'X',
-      destinationDuration: '60',
-      transportation: 'Car',
-      distance: '3km',
-      estimatedPrice: '15 AUD',
-      startLocation: '48.8566,2.3522',
-      endLocation: '48.8606,2.3376',
-      photo_reference: 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx',
-      rating: 4.5,
-      user_ratings_total: 2000
-    },
-  ],
-  2: [
-    {
-      date: '2024-09-29',
-      time: '8:00',
-      duration: '10',
-      destination: 'X',
-      destinationDescrib: 'X',
-      destinationDuration: '60',
-      transportation: 'Car',
-      distance: '3km',
-      estimatedPrice: '15 AUD',
-      startLocation: '48.8566,2.3522',
-      endLocation: '48.8606,2.3376',
-      photo_reference: 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx',
-      rating: 4.5,
-      user_ratings_total: 2000
-    },
-  ],
+    }
+  ]
 }
 
+function calculate_radius(TimeSpendMin:number, TimeSpendMax:number, transportation: string){
+  let minRadius = 0
+  let maxRadius = 0
+  switch (transportation) {
+    case 'driving':
+      minRadius = TimeSpendMin * 300
+      maxRadius = TimeSpendMax * 300
+      break
+    case 'walking':
+      minRadius = TimeSpendMin * 60
+      maxRadius = TimeSpendMax * 60
+      break
+    case 'transit':
+      minRadius = TimeSpendMin * 150
+      maxRadius = TimeSpendMax * 150
+      break
+  }
+  return [minRadius, maxRadius]
+}
 // Function to generate meal plan for the day or the next day, including breakfast, lunch, dinner
 export async function generatePlan_restaurant(
   gps_location: string,
@@ -165,33 +85,12 @@ export async function generatePlan_restaurant(
   const departureTime = new Date(
     now.getTime() - now.getTimezoneOffset() * 60000
   ).toISOString()
-  console.log(departureTime)
   let currentLocation = gps_location
-  const time = departureTime.slice(11, 19)
-  const hours = parseInt(time.slice(0, 2))
-  const minutes = parseInt(time.slice(3, 5))
-  console.log(`Current time: ${hours}:${minutes}`)
   let numMeals = 1
-  let minRadius = 0
-  let maxRadius = 0
-  switch (travel_mode) {
-    case 'driving':
-      minRadius = TimeSpendMin * 350
-      maxRadius = TimeSpendMax * 350
-      break
-    case 'walking':
-      minRadius = TimeSpendMin * 84
-      maxRadius = TimeSpendMax * 84
-      break
-    case 'transit':
-      minRadius = TimeSpendMin * 200
-      maxRadius = TimeSpendMax * 200
-      break
-  }
+  const [minRadius, maxRadius] = calculate_radius(TimeSpendMin, TimeSpendMax, travel_mode)
   try {
     // Get gpt generate history
     const history = await getHistory()
-    // console.log(`History: ${JSON.stringify(history)}`)
     let planJson: Plan = { 1: [] }
     for (let i = 0; i < numMeals; i++) {
       const placesJson = await getNearbyPlaces(
@@ -202,7 +101,6 @@ export async function generatePlan_restaurant(
         maxPrice
       )
       let filteredPlacesJson = filterGoogleMapDataByMinRadius(filterGoogleMapData(placesJson, currentLocation), minRadius)
-      // console.log(filteredPlacesJson);
 
       const locations: string[] = []
       filteredPlacesJson.forEach((place) => {
@@ -216,7 +114,6 @@ export async function generatePlan_restaurant(
         travel_mode
       )
       const filteredDistanceMatrix = filterDistanceMatrixData(distanceMatrix)
-      // console.log(JSON.stringify(filteredDistanceMatrix))
       // Let GPT to generate the next activity
       if (!GPT_KEY) {
         console.error('GPT_KEY is not defined.')
@@ -224,11 +121,7 @@ export async function generatePlan_restaurant(
       }
 
       const url = 'https://api.openai.com/v1/chat/completions'
-      // const requestMessage = `It's ${currentTime} now. If the current time exceed 9 pm, Give me the meal plan for next day. Please fill in the "transportation" with ${travel_mode || "driving"} (Capitalize the first letter).There are ${numMeals-i} meals remaining.
-      // If 3 meals remaining, give me the breakfast plan, "time" must be a reasonable breakfast time. If 2 meals remaining, give me the lunch plan, "time" must be a reasonable lunch time. If 1 meal remaining, give me dinner plan,"time" must be a reasonable dinner time.
-      // All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
-      // The current plan is ${JSON.stringify(planJson)}. [Important] 1.Don't let me go to the same attraction twice.2.The "time" of returned activity should be later than the "time"+"destinationDuration"+"duration" of last activity in the current plan.`
-      const requestMessage = `It's ${TimeSpendMax} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter). Let the "time" of the plan be the current time.
+      const requestMessage = `It's ${departureTime} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter). Let the "time" of the plan be the current time.
             All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}. You don't need to change the photo_reference from the given data, just use the photo_reference in the given data.
             The current plan is ${JSON.stringify(planJson)}. The history is ${JSON.stringify(history)}. "title" in history is the names of places."visit_count" is the times the user has visited this place.`
 
@@ -294,27 +187,8 @@ export async function generatePlan_cafe(
   const departureTime = new Date(
     now.getTime() - now.getTimezoneOffset() * 60000
   ).toISOString()
-  console.log(departureTime)
-  const time = departureTime.slice(11, 19)
-  const hours = parseInt(time.slice(0, 2))
-  const minutes = parseInt(time.slice(3, 5))
-  console.log(`Current time: ${hours}:${minutes}`)
-  let minRadius = 0
-  let maxRadius = 0
-  switch (travel_mode) {
-    case 'driving':
-      minRadius = TimeSpendMin * 300
-      maxRadius = TimeSpendMax * 300
-      break
-    case 'walking':
-      minRadius = TimeSpendMin * 60
-      maxRadius = TimeSpendMax * 60
-      break
-    case 'transit':
-      minRadius = TimeSpendMin * 150
-      maxRadius = TimeSpendMax * 150
-      break
-  }
+
+  const [minRadius, maxRadius] = calculate_radius(TimeSpendMin, TimeSpendMax, travel_mode)
   try {
     const history = await getHistory()
     let planJson: Plan = { 1: [] }
@@ -327,7 +201,6 @@ export async function generatePlan_cafe(
         maxPrice
       )
       let filteredPlacesJson = filterGoogleMapDataByMinRadius(filterGoogleMapData(placesJson, currentLocation), minRadius)
-      // console.log(filteredPlacesJson);
       const locations: string[] = []
       filteredPlacesJson.forEach((place) => {
         const location = place.vicinity
@@ -340,7 +213,6 @@ export async function generatePlan_cafe(
         travel_mode
       )
       const filteredDistanceMatrix = filterDistanceMatrixData(distanceMatrix)
-      // console.log(JSON.stringify(filteredDistanceMatrix))
       // Let GPT to generate the next activity
       if (!GPT_KEY) {
         console.error('GPT_KEY is not defined.')
@@ -348,11 +220,7 @@ export async function generatePlan_cafe(
       }
 
       const url = 'https://api.openai.com/v1/chat/completions'
-      // const requestMessage = `It's ${currentTime} now. If the current time exceed 5 pm, Give me the plan to go to a cafe next day.  Please fill in the "transportation" with ${travel_mode || "driving"} (Capitalize the first letter).There are ${numMeals-i} cafe remaining to go.
-      // If there are 2 cafe to go, give me a plan to go a cafe in the morning, "time" must be a reasonable breakfast time. If there is 1 cafe to go, give me a plan to go a cafe in the afternoon, "time" must be a reasonable afternoon tea time.
-      // All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
-      // The current plan is ${JSON.stringify(planJson)}. [Important] 1.Don't let me go to the same attraction twice.2.The "time" of returned activity should be later than the "time"+"destinationDuration"+"duration" of last activity in the current plan.`
-      const requestMessage = `It's ${TimeSpendMax} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).Let the "time" of the plan be the current time.
+      const requestMessage = `It's ${departureTime} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).Let the "time" of the plan be the current time.
             All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
             The current plan is ${JSON.stringify(planJson)}.The history is ${JSON.stringify(history)}. "title" in history is the names of places."visit_count" is the times the user has visited this place.`
       const requestBody = {
@@ -416,28 +284,8 @@ export async function generatePlan_entertainment(
   const departureTime = new Date(
     now.getTime() - now.getTimezoneOffset() * 60000
   ).toISOString()
-  console.log(departureTime)
-  const time = departureTime.slice(11, 19)
-  const hours = parseInt(time.slice(0, 2))
-  const minutes = parseInt(time.slice(3, 5))
-  console.log(`Current time: ${hours}:${minutes}`)
 
-  let minRadius = 0
-  let maxRadius = 0
-  switch (travel_mode) {
-    case 'driving':
-      minRadius = TimeSpendMin * 300
-      maxRadius = TimeSpendMax * 300
-      break
-    case 'walking':
-      minRadius = TimeSpendMin * 60
-      maxRadius = TimeSpendMax * 60
-      break
-    case 'transit':
-      minRadius = TimeSpendMin * 150
-      maxRadius = TimeSpendMax * 150
-      break
-  }
+  const [minRadius, maxRadius] = calculate_radius(TimeSpendMin, TimeSpendMax, travel_mode)
   try {
     const history = await getHistory()
     console.log(`History: ${JSON.stringify(history)}`)
@@ -463,7 +311,6 @@ export async function generatePlan_entertainment(
         travel_mode
       )
       const filteredDistanceMatrix = filterDistanceMatrixData(distanceMatrix)
-      // console.log(JSON.stringify(filteredDistanceMatrix))
       // Let GPT to generate the next activity
       if (!GPT_KEY) {
         console.error('GPT_KEY is not defined.')
@@ -471,11 +318,7 @@ export async function generatePlan_entertainment(
       }
 
       const url = 'https://api.openai.com/v1/chat/completions'
-      // const requestMessage = `It's ${currentTime} now. If the current time exceed 5 pm, Give me the plan to go to a cafe next day.  Please fill in the "transportation" with ${travel_mode || "driving"} (Capitalize the first letter).There are ${numMeals-i} cafe remaining to go.
-      // If there are 2 cafe to go, give me a plan to go a cafe in the morning, "time" must be a reasonable breakfast time. If there is 1 cafe to go, give me a plan to go a cafe in the afternoon, "time" must be a reasonable afternoon tea time.
-      // All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
-      // The current plan is ${JSON.stringify(planJson)}. [Important] 1.Don't let me go to the same attraction twice.2.The "time" of returned activity should be later than the "time"+"destinationDuration"+"duration" of last activity in the current plan.`
-      const requestMessage = `It's ${TimeSpendMax} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).Let the "time" of the plan be the current time.
+      const requestMessage = `It's ${departureTime} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).Let the "time" of the plan be the current time.
             All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
             The current plan is ${JSON.stringify(planJson)}.The history is ${JSON.stringify(history)}. "title" in history is the names of places."visit_count" is the times the user has visited this place.`
       const requestBody = {
@@ -538,28 +381,8 @@ export async function generatePlan_milktea(
   const departureTime = new Date(
     now.getTime() - now.getTimezoneOffset() * 60000
   ).toISOString()
-  console.log(departureTime)
-  const time = departureTime.slice(11, 19)
-  const hours = parseInt(time.slice(0, 2))
-  const minutes = parseInt(time.slice(3, 5))
-  console.log(`Current time: ${hours}:${minutes}`)
 
-  let minRadius = 0
-  let maxRadius = 0
-  switch (travel_mode) {
-    case 'driving':
-      minRadius = TimeSpendMin * 300
-      maxRadius = TimeSpendMax * 300
-      break
-    case 'walking':
-      minRadius = TimeSpendMin * 60
-      maxRadius = TimeSpendMax * 60
-      break
-    case 'transit':
-      minRadius = TimeSpendMin * 150
-      maxRadius = TimeSpendMax * 150
-      break
-  }
+  const [minRadius, maxRadius] = calculate_radius(TimeSpendMin, TimeSpendMax, travel_mode)
     try {
       const history = await getHistory()
       console.log(`History: ${JSON.stringify(history)}`)
@@ -591,11 +414,7 @@ export async function generatePlan_milktea(
         }
   
         const url = 'https://api.openai.com/v1/chat/completions'
-        // const requestMessage = `It's ${currentTime} now. If the current time exceed 5 pm, Give me the plan to go to a cafe next day.  Please fill in the "transportation" with ${travel_mode || "driving"} (Capitalize the first letter).There are ${numMeals-i} cafe remaining to go.
-        // If there are 2 cafe to go, give me a plan to go a cafe in the morning, "time" must be a reasonable breakfast time. If there is 1 cafe to go, give me a plan to go a cafe in the afternoon, "time" must be a reasonable afternoon tea time.
-        // All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
-        // The current plan is ${JSON.stringify(planJson)}. [Important] 1.Don't let me go to the same attraction twice.2.The "time" of returned activity should be later than the "time"+"destinationDuration"+"duration" of last activity in the current plan.`
-        const requestMessage = `It's ${TimeSpendMax} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).Let the "time" of the plan be the current time.
+        const requestMessage = `It's ${departureTime} now. Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).Let the "time" of the plan be the current time.
               All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
               The current plan is ${JSON.stringify(planJson)}.The history is ${JSON.stringify(history)}. "title" in history is the names of places."visit_count" is the times the user has visited this place.`
         const requestBody = {
@@ -672,22 +491,7 @@ export async function generatePlan_attractions(
     numAttractions = 5
   }
 
-  let minRadius = 0
-  let maxRadius = 0
-  switch (travel_mode) {
-    case 'driving':
-      minRadius = TimeSpendMin * 300
-      maxRadius = TimeSpendMax * 300
-      break
-    case 'walking':
-      minRadius = TimeSpendMin * 60
-      maxRadius = TimeSpendMax * 60
-      break
-    case 'transit':
-      minRadius = TimeSpendMin * 150
-      maxRadius = TimeSpendMax * 150
-      break
-  }
+  const [minRadius, maxRadius] = calculate_radius(TimeSpendMin, TimeSpendMax, travel_mode)
   try {
     const history = await getHistory()
     // console.log(`History: ${JSON.stringify(history)}`)
@@ -723,7 +527,7 @@ export async function generatePlan_attractions(
       }
 
       const url = 'https://api.openai.com/v1/chat/completions'
-      const requestMessage = `It's ${TimeSpendMax} now. If the current time exceed 22 pm, Give me the plan to go to tourist attractions next day.  Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).There are ${numAttractions - i} tourist attractions remaining to go.
+      const requestMessage = `It's ${departureTime} now. If the current time exceed 22 pm, Give me the plan to go to tourist attractions next day.  Please fill in the "transportation" with ${travel_mode || 'driving'} (Capitalize the first letter).There are ${numAttractions - i} tourist attractions remaining to go.
             All the places around: ${JSON.stringify(filteredPlacesJson)}. All distances and durations from current location to the places one by one in previous data: ${JSON.stringify(filteredDistanceMatrix[0])}.
             The current plan is ${JSON.stringify(planJson)}.The history is ${JSON.stringify(history)}. "title" in history is the names of places."visit_count" is the times the user has visited this place.`
       const requestBody = {
@@ -783,47 +587,4 @@ function filterGoogleMapDataByMinRadius(filteredPlacesJson: any[], minRadius:num
   return filteredPlacesJson
 }
 
-// export async function askAboutPlan(
-//   question: string,
-//   plan: Plan
-// ): Promise<string | void> {
-//   if (!GPT_KEY) {
-//     console.error('GPT_KEY is not defined.')
-//     return
-//   }
 
-//   const url = 'https://api.openai.com/v1/chat/completions'
-
-//   const requestMessage = `User Question: ${question}. The current plan: ${plan}`
-//   const requestBody = {
-//     model: 'gpt-4o-mini',
-//     messages: [
-//       {
-//         role: 'system',
-//         content: `You are a robot to answer questions about the plan. You only need to reply in one or two sentences in text not in json.
-//                 "destination" is the true name of the destination in the data given."time" is the recommended start time to go to the destination."date" is the date of the activity."destination describ" is the description of the destination. 
-//                 "destination duration" is the recommended time staying at the destination in minutes."estimated price" is the estimated money spent in this destination (estimate according to the price level in the given data)."startLocation" and "endLocation" are location in latitude and longitude.
-//                 Do not return anything beyond the given plan data.`,
-//       },
-//       { role: 'user', content: requestMessage },
-//     ],
-//     max_tokens: 1000,
-//   }
-
-//   try {
-//     const response = await fetch(url, {
-//       method: 'POST',
-//       headers: {
-//         'Content-Type': 'application/json',
-//         Authorization: `Bearer ${GPT_KEY}`,
-//       },
-//       body: JSON.stringify(requestBody),
-//     })
-
-//     const data = await response.json()
-//     console.log(data.choices[0].message.content)
-//     return data.choices[0].message.content
-//   } catch (error) {
-//     console.error('Error calling GPT API:', error)
-//   }
-// }
